PIM-7182: Fix clean mongodb command
An issue has been raised whenever a user would use the CleanMongoDBCommand
with products having reference data.

This command is used to clean the mongoDB when some data (families /
attributes) are not in sync with the Mysql DB (ie, entities have been
deleted from mysql but not in mongodb).

**Problem:**
The command would remove the values of reference data attributes even
when the selected attribute option still existed in mysql.

**Solution:**
This error was due to wrong check in the command. This patch fixes it.